requirements.txt
alembic==1.6.2
appdirs==1.4.4
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
chardet==4.0.0
cleo==0.8.1
click==8.0.0
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
decorator==5.0.9
distlib==0.3.1
dnspython==2.1.0
docopt==0.6.2
dominate==2.6.0
email-validator==1.1.2
filelock==3.0.12
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
infinity==1.5
intervals==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.0
keyring==21.8.0
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==2.0.0
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pipreqs==0.4.10
pkginfo==1.7.0
poetry==1.1.6
poetry-core==1.0.3
psycopg2==2.8.6
ptyprocess==0.7.0
pylev==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pywin32-ctypes==0.2.0
requests==2.25.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.15
tomlkit==0.7.2
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.6
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-Components==0.10.5
yarg==0.1.9
